deployment/hrdlbackend/requirements.txt
Pillow
Django==3.2.7
django-debug-toolbar==2.2
gunicorn==20.0.4
psycopg2-binary
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.1.0
pillow
tqdm
numpy
matplotlib
python-dateutil
pandas==1.3.5
torch
scikit-learn==1.0.1
python-binance
django-cors-headers
django-extensions
djangorestframework-simplejwt
django-celery-beat
celery==4.4.2
redis>=3.2.0
pymysql
jaraco.docker
ccxt
pyyaml
coinpaprika-client
pycurl